Quick Fix: ParticleTextView Rounding
*Int rounding results in incorrect positioning. Fix by performing arithmetic in float and then cast to int.

+++ particletextview/src/main/java/com/yasic/library/particletextview/View/ParticleTextView.java
@@ -110,8 +110,8 @@ public class ParticleTextView extends View {
     }
 
     private int[][] bitmapTransition() {
-        int centerX = (getWidth() / 100) * xPosition;
-        int centerY = (getHeight() / 100) * yPosition;
+        int centerX = (int) (((float) getWidth() / 100) * xPosition);
+        int centerY = (int) (((float) getHeight() / 100) * yPosition);
         int[][] colorArray;
         textPaint = initTextPaint();
         bitmap = Bitmap.createBitmap(getWidth(), getHeight(), Bitmap.Config.ARGB_8888);
